requirements.txt
pygame==2.6.1
Pillow==11.0.0
